dipa2_dataset/requirements.txt
matplotlib==3.5.3
numpy==1.23.5
pandas==1.5.3
Pillow==10.0.0
pytorch_lightning==1.9.1
scikit_learn==1.3.0
torch==1.13.1
torchmetrics==0.11.1
torchvision==0.14.1
wandb==0.13.10
